(maint) Remove ext/rack/README reference from debian/rules
In a previous commit, files in ext/rack/files were moved up a directory to
ext/rack and some unneeded files were removed. Among the removed files was
README, which was unfortunately still used in the debian/rules file as a
supplemental doc for the puppetmaster-passenger package. The lack of
ext/rack/README caused the rules evaluation to fail, which caused debian
packaging to fail. This commit addresses that by removing the README references
from the debian/rules file.