(maint) Rename expected apache vhost in debian/rules
In a previous commit, files in ext/rack/files were moved up a directory to
ext/rack and some files were renamed. Among the renamed files was
apache2.conf, which was renamed to example-passenger-vhost.conf.  This caused
the rules evaluation to fail, which caused debian packaging to fail. This
commit addresses that by renaming the vhost references in the debian/rules
file.