(#22363) Remove support for compound assignments
Compund (deflating) assignments where not completely implemented
and it turned out to be a bad idea (i.e. indirect variable
creating where [a,b,c] = [1,2,3] produces a = 1, b = 2, c = 3,
and [a,b,c] = { x=>10, a=>1, b=>2, c =>3} produces the same).

Not much is saved in terms of typing while having poor readability.
It also adds further things that can not be statically analyzed
since variable names can be dynamically computed. This can be abused.

The construct was removed instead of completed.

This also adds assignment tests for =, += and -=