(#22363) Fix problems in TypeParser
The TypeParser misinterpreted several types as being ResourceType
if an attempt was made to parameterize a non parameterized type.

The type Object is now also supported.

When the TypeParser was used to only interpret it failed when
there was an error - this because the type parser was not used
to parse the original string. All ast objects have a Locator when
they come from the future parser and it holds the parsed string.
Now, the string is obtained from the locator if not originally
parsed via the type parser. (This should be cleaned up later).