(#22363) Make test for [] on Class to get parameter really work
The test was botched, it needed to do an import_ast to get class
initialized, and then evaluate its code, not the class itself.

A test in access_ops no longer works (the semantics of [] changed) and
access_ops is not instrumented to do full code import. (The test
is now in evaluating_parser_spec instead).

The e4_parser_adapter was to eager to report a .rb file as being
not supported (it should have checked if it was in "string mode" before
failing).

PopsBridge had a typo.