(#22363) Add strict variable lookup mode
This adds the setting :strict_variables. When used the scope
will throw :unkown_variable (which if not handled, will raise
an exception). The future evaluator can handle this, but none of
the other parsers/evaluator options. Testing of the future evaluator
turns on strict variables mode (and has tests that an error is raised).

This commit also contains maintenance (removal of unused code), fixing
of the pops bridge (special case when evaluating without a proper boot)