(#23236) Deprecate Puppet::Util::SUIDManager.run_and_capture
This deprecates run_and_capture from SUIDManager for
Puppet::Util::Execution.execute(). Run and capture is returning $CHILDSTATUS
where execute returns an exit status as part of output. This means that run and
capture could have potential issues when being used, especially with Windows
which is known to have issues due to the truncation of exit codes greater than
256.