(maint) Reset hiera in test_helper to avoid order issues in specs
Andrew Parker ran into spec order issues when merging into pe-puppet
which came down to the hiera instance needing to be reset in the
test_helper.  We are back-porting these changes into puppet so that we
are insulated from the same issues and to keep parity.

pe-puppet/3.2.x: 5ba1392 and 14b36ef