(#18931) File source_permissions setting applies to local files
The new source_permissions setting was original excluded from effecting
files with a local source (where the file is a path on the agent).  This
leaves a problem on Windows, because attempting to copy mode, even from
a local file, still produces insufficient acls.  This patch allows
source_permissions to affect local files, so that mode can be ignored,
and deprecation warnings appear.  We also added a debug message to the
resource to make it possible to pinpoint which File resources are using
this deprecated method.