(maint) Extract common acceptance setup script for Jenkins
We have six Jenkins jobs which currently run acceptance, and will soon
have another four for per commit tests.  There is a shell script used to
prepare each job for Rake execution of the actual test run.  This script
currently exists in every job and is slightly divergent across jobs, and
makes updating the prep for each job time consuming and error prone.  So
I'm pulling it into acceptance as a single script to handle untarring of
the artifact and bundle set up for any acceptance job.

This change also provides a means for specifying the git daemon along
with the fork so that you can install from local commits.

Finally it moves repo_proxy and add_el_extras beaker settings into the
common options.