(#22363) Fix problems with undef equality & problems from added tests
Tests from the parser integration tests were added to the unit tests
for the new evaluator (they were not very good integration tests). The
old test ran for both "current" and "future" and now only run for
"current". The relevant tests were moved to the evaluator unit tests.

This showed several problems in the PopsBridge (typos) for yet
unexercised code paths. These are now fixed.

As the integration/parser/compiler_spec tests were being investigated
to run also for the new parser, problems were found with comparison
against literal undef. Tests were added for this in the evaluator unit
test and the problem corrected (undef compared as :undef == nil => false
which is not the intention. Comparsions :undef == nil, and nil == :undef
are now done correctly.

Exceptions did not show up correctly when file is empty string. This
also fixes the logic where it is checked if a file is known (nil and
empty string are now considered to be "file unknown" instead of file ""
known.