(#22363) Make compilation work using new evaluator
An extract_type_and_title method was added to the type calculator.
It extract method extracts type/title from catalog types. This is used
in the 3x interface as 3x cannot directly deal with types.

Class methods for the API methods are also added to make the type
calculator easier to use (no need to keep an instance around; although
doing so is slightly faster).

Tests used references like Class[Baz] where Baz is a PResourceType. This
is now supported by the AccessOperator, but it is questionable if this
should be supported.

The runtime3 support now has an initializer that is called from the
evaluator implementation. It creates a value converter that is needed to
convert the runtime objects from 4x to 3x (types, and regular
expressions) that are not supported as first class objects in the 3x
runtime.

The model label provider now also provides labels for the type model.
There are additional validation / runtime issues.

The validating parser was too fond of one particular acceptor and errors
from previous runs were accumulated.

The evaluating parser did not clear the definitions and name stack
arrays which led to accumulation of classes, definitions and nodes in
subsequent use of the same instance of evaluating parser.