(#22363) Remove support for extract type and title from type calculator
The support for 3x logic to convert catalog types to name and title
array was intended for calls from within 3x logic. This was not a
desired solution and the the support in the type calculator ended up
being unused. This removes it. (The corresponding logic is instead found
in the runtime3_support module).