(#23141) Fix remount on bsd if options are specified
The initial fix for 23141 didn't account for a remount when
options were specified.  This caused integration tests to
fail because a) the integration tests assumed mounted is always
false when the mountcmd is called (not true for a remount), and
b) the integration tests weren't updated to consider the 'update'
option.

This patch updates the remount method to append ',update' if
there are already options, and updates the integration tests
accordingly.