(maint) Skip enc test on lucid
Currently the lucid runs of this test are failing pretty frequently because
of difficulties getting to rubyforge.org.  This change would skip this
test on lucid.

We're already skipping this test on other platforms due to various difficulties
(read the log for this file). And activerecord is a deprecated feature, so this
test won't be around for too much longer.  Meanwhile, we have decent coverage
of this feature since it still runs against RHEL5 and newer Ubuntus.