(#11383) support purging of all cronjobs
Add a method to the Util::FileType class that lists all accessible
crontab files stored in the appropriate location. Using this method,
have the crontab provider add the present files as targets.

Consequently, the crontab provider can now generate resources for all
crontab entries in the system via its instances method. This is all
that is required to allow purging of all those that are not managed.