(#11383) update crontab provider unit tests
Two tests need changing, because when running rspec as root, it should
actually be expected that there are cron resources, even though the
target user is absent or has no crontab. Change the test semantics so that
cron entries of other users are ignored. This fix was proposed by
John Bollinger.

Due to the parsing of system crontabs in addition to the fixture
(again, only if running a root), the target_object method will be
invoked more than once. We can no longer rely on expect(), because
we cannot know the number of invocations on any developer's machine.
Use stubs() instead and ignore the number of invocations. Technically,
this removes a check, but it's not an important aspect of the tests
in question.