(yardoc) Fix stale yardoc for Resource [] operator
The documentation had typos and was wrong since support was added
for getting the property of a fully qualified resource reference.
I.e. File['foo']['bar'] gets the 'bar' property of the referenced file
resource. 
Also had stale documentation stating that [] keys produced a deep copy.
That is wrong - it produces a syntax error, and if done from code, it
fails with an arity error.