(#22363) Remove validation of relationship expressions
Relationship expressions had validation that their LHS and RHS where
of certain types. This was moot since almost all expressions are capable
of producing a value that it is meaningful to use in a relationship
expression. The restrictions were artificial due to problems in the 3x
grammar.

This also adds a test for relation between two resources.

More work is required to complete the testing for relationships and
resource expression R-values.