(#23373) Always open config so we are reading from the beginning
Not all operating systems start reading from the same position with 'a+'
when opening the file.  This ensures that we create the file with
default permissions, and begin reading from the first line.

We also make some acceptance changes for testing puppet config set, and
remove an acceptance test which was of dubious value.