(PUP-1276) Normalize tidy paths
Previously, a tidy resource could create a duplicate file resource due to a
mismatch between back and forward slashes on Windows.

This commit munges the tidy path parameter using File.expand_path, so on
Windows, `C:\dir` becomes `C:/dir`. When that is File.join'ed with a file
needing to be tidy'ed, the resulting path contains only forward slashes,
which is consistent with the path parameter for the file type.