Maint: Move pkgdmg documentation into provider description; document changed behavior
This wiki page was more longwinded than it needed to be; you can really sum it
up in about four bullet points. So I did that, and removed the wiki link.

Also, while investigating what the wiki said, it became clear that this provider
no longer actually uses open-uri: a Kernel.open call was changed to File.open,
but apparently nobody was using file:/// URLs (or anything weirder), so it went
unnoticed. So I added some comments for future maintainers of this code.